+++ src/components/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod applicability;
 pub(crate) mod ring_topology;
 pub(crate) mod size_topology;
+pub(crate) mod stereochemical_burden;